move module init+shutdown logic into module.rs
